attempt to stop blocks from spilling offscreen in handheld view
And to keep the blocks width matching the contents, but I failed to do
that, the styles get a bit complicated.  It's acceptable, but I may
revisit this later.  The problem is visible when viewing the page in a
small window, like a phone.  The blocks are narrower than the content.
They should match.  However, the responsive stylesheets apply a 100% o
the second sidebard blocks which makes it expand too much; then there
is something else (I forget what) which makes the blocks too narrow.